cleanup routes from app.js and make it look better
