+++ app.js
@@ -13,43 +13,10 @@ var port= process.env.PORT || 3000;
 app.use(bodyParser.urlencoded({extended:true}));
 app.use(bodyParser.json());
 
-var bookRouter=express.Router();
+bookRouter=require('./Routes/bookRoutes')(Book);
 
-bookRouter.route('/Books')
-    .post(function (req,res) {
-        var book=new Book(req.body);
-        book.save();
-        res.status(201).send(book);
-
-    })
-    .get(function (req, res) {
-
-        var query={};
-
-        if(req.query.genre){
-            query.genre=req.query.genre;
-        }
-       //var responseJson={hello: "This is my API"};
-        Book.find(query, function (err, books) {
-            if(err)
-                res.status(500).send(err);
-            else
-                res.json(books);
-        });
-    });
-
-bookRouter.route('/Books/:bookId')
-    .get(function (req, res) {
-
-        Book.findById(req.params.bookId, function (err, book) {
-            if(err)
-                res.status(500).send(err);
-            else
-                res.json(book);
-        });
-    });
-app.use('/api', bookRouter);
-
+app.use('/api/books', bookRouter);
+app.use('/api/authors', bookRouter);
 
 
 app.get('/', function (req, res) {
